Move the includes analyzer code into the src directory
Keeping the structure consistent to how the aspect and automatic fixes
code is partitioned.
