+++ src/analyze_includes/test/get_dependencies_test.py
@@ -1,7 +1,7 @@
 import unittest
 from pathlib import Path
 
-from src.get_dependencies import (
+from src.analyze_includes.get_dependencies import (
     AvailableInclude,
     IncludeUsage,
     get_available_dependencies,
@@ -64,7 +64,7 @@ class TestAvailableInclude(unittest.TestCase):
 
 class TestGetAvailableDependencies(unittest.TestCase):
     def test_load_full_file(self):
-        deps = get_available_dependencies(Path("src/test/data/deps_info_full.json"))
+        deps = get_available_dependencies(Path("src/analyze_includes/test/data/deps_info_full.json"))
 
         self.assertEqual(len(deps.private), 2)
         self.assertEqual(len(deps.public), 2)
@@ -92,7 +92,7 @@ class TestGetAvailableDependencies(unittest.TestCase):
         self.assertEqual(deps.self.hdrs, [AvailableInclude("self/a.h"), AvailableInclude("self/b.h")])
 
     def test_load_empty_file(self):
-        deps = get_available_dependencies(Path("src/test/data/deps_info_empty.json"))
+        deps = get_available_dependencies(Path("src/analyze_includes/test/data/deps_info_empty.json"))
 
         self.assertEqual(len(deps.private), 0)
         self.assertEqual(len(deps.public), 0)
