+++ src/analyze_includes/test/parse_config_test.py
@@ -1,18 +1,18 @@
 import unittest
 from pathlib import Path
 
-from src.parse_config import load_config
+from src.analyze_includes.parse_config import load_config
 
 
 class TestLoadConfig(unittest.TestCase):
     def test_load_empty(self):
-        ignored_inclues, extra_ignoredf_includes = load_config(Path("src/test/data/config_empty.json"))
+        ignored_inclues, extra_ignoredf_includes = load_config(Path("src/analyze_includes/test/data/config_empty.json"))
 
         self.assertEqual(ignored_inclues, [])
         self.assertEqual(extra_ignoredf_includes, [])
 
     def test_load_config(self):
-        ignored_includes, extra_ignored_includes = load_config(Path("src/test/data/config.json"))
+        ignored_includes, extra_ignored_includes = load_config(Path("src/analyze_includes/test/data/config.json"))
 
         self.assertEqual(len(ignored_includes), 2)
         self.assertTrue("foo" in ignored_includes)
